packages/sdk/pyproject.toml
[project]
name = "safer-sdk"
version = "0.1.3"
description = "SAFER SDK — instrument any AI agent framework with one line. Hooks for Claude SDK, OpenAI, OpenAI Agents, LangChain, Google ADK, AWS Strands, AWS Bedrock, and CrewAI."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Mert Ozbas", email = "mertozbas@gmail.com" },
    { name = "SAFER Contributors" },
]
maintainers = [
    { name = "Mert Ozbas", email = "mertozbas@gmail.com" },
]
keywords = [
    "ai",
    "agents",
    "agent-framework",
    "observability",
    "instrumentation",
    "security",
    "llm",
    "claude",
    "anthropic",
    "openai",
    "langchain",
    "strands",
    "google-adk",
    "bedrock",
    "crewai",
    "hooks",
    "tracing",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.5",
    "httpx>=0.27",
    "websockets>=12.0",
    "anyio>=4.3",
]

[project.urls]
Homepage = "https://github.com/hashtagemy/safer"
Documentation = "https://github.com/hashtagemy/safer#readme"
Repository = "https://github.com/hashtagemy/safer"
Issues = "https://github.com/hashtagemy/safer/issues"
Changelog = "https://github.com/hashtagemy/safer/releases"

[project.optional-dependencies]
claude = ["anthropic>=0.40"]
langchain = ["langchain>=0.2", "langchain-core>=0.2"]
openai = ["openai>=1.40"]
openai-agents = ["openai-agents>=0.0.5", "openai>=1.40"]
google-adk = ["google-adk>=1.0"]
strands = ["strands-agents>=1.0"]
bedrock = ["boto3>=1.34"]
crewai = ["crewai>=0.55"]
otel = [
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp-proto-http>=1.28",
]
otel-anthropic = [
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp-proto-http>=1.28",
    "opentelemetry-instrumentation-anthropic>=0.45b0",
]
otel-openai = [
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp-proto-http>=1.28",
    "opentelemetry-instrumentation-openai>=0.45b0",
]
all = [
    "anthropic>=0.40",
    "langchain>=0.2",
    "langchain-core>=0.2",
    "openai>=1.40",
    "openai-agents>=0.0.5",
    "google-adk>=1.0",
    "strands-agents>=1.0",
    "boto3>=1.34",
    "crewai>=0.55",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp-proto-http>=1.28",
]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/safer"]
